@brandon-rhodes 2019-04-12
Thanks for thinking of ways to try to make the library faster! Could the problem be that the `if any(...)` approach means that an adjustment that normally would only be made for one of the satellites matching the `if` condition will now be made to all of them, throwing the other ones off?