@dsholes 2019-04-12
Hi Brandon,

I found a bug. I'm using `np.where()` for the if any() cases, but for the `em` line, I switched the order of the arguments:

`em = np.where(em < 1.0e-6, em, 1.0e-6);` should be
`em = np.where(em < 1.0e-6, 1.0e-6, em);`

That actually fixed everything. Now the results agree exactly for my tests. 

I realized however that I edited an old version of `sgp4.propagation`, so I've re-forked, deleted my pull request, and I'll make a new one after I've finished the edits.

I imagine more work will still need to be done for when `sat.method=='d'`. I've only touched `propagation.sgp4()`